Correction to isLoaded check in pushTag
Version 1.3.0 converted addGtmToDom into an asynchronous function; however, the resulting change to the logic in pushTag means that if isLoaded is false, both calls of pushOnDataLayer within the pushTag method are called (the inner "return resolve()" does not terminate the function because it is postponed by the addGtmToDom execution). This leads to the first item of the session being pushed twice.

This pull request alters the second half of pushTag to only be called if isLoaded is false.